requirements.txt
# Core app
Flask==3.1.2
openai==2.30.0
python-dotenv==1.2.2

# Web ingestion
requests==2.32.3
beautifulsoup4==4.14.2
urllib3==2.3.0

# Document parsing
pandas==2.3.3
PyPDF2==3.0.1
python-docx==1.2.0

# Google Drive ingestion
google-api-python-client==2.176.0
google-auth==2.49.1
google-auth-oauthlib==1.2.2
